write to multiple mcu
