+++ src/rest_api/actions/read_info_action.py
@@ -92,7 +92,7 @@ class DoorsToJSON():
             "BeltCard": data[4],
             "WindowStatus": data[5],
         }
-        return (response_to_frontend)
+        return response_to_frontend
 
 
 class ReadInfo(Action):
@@ -244,39 +244,39 @@ class ReadInfo(Action):
             return e.message
 
 
-def read_from_doors(self):
-    id_door = self.id_ecu[1]
-    id = self.my_id * 0x100 + id_door
+    def read_from_doors(self):
+        id_door = self.id_ecu[1]
+        id = self.my_id * 0x100 + id_door
 
-    try:
-        log_info_message(logger, "Changing session to default")
-        self.generate.session_control(id, 0x01)
-        self._passive_response(SESSION_CONTROL, "Error changing session control")
-        log_info_message(logger, "Changing session to default")
-        self.generate.session_control(id, 0x01)
-        self._passive_response(SESSION_CONTROL, "Error changing session control")
-        self._authentication(id)
+        try:
+            log_info_message(logger, "Changing session to default")
+            self.generate.session_control(id, 0x01)
+            self._passive_response(SESSION_CONTROL, "Error changing session control")
+            log_info_message(logger, "Changing session to default")
+            self.generate.session_control(id, 0x01)
+            self._passive_response(SESSION_CONTROL, "Error changing session control")
+            self._authentication(id)
 
-        log_info_message(logger, "Reading data from doors")
+            log_info_message(logger, "Reading data from doors")
 
-        Door = self._read_by_identifier(id, IDENTIFIER_DOOR)
-        serial_number = self._read_by_identifier(id, IDENTIFIER_DOOR_SERIALNUMBER)
-        cigarette_lighter_voltage = self._read_by_identifier(id, IDENTIFIER_LIGHTER_VOLTAGE)
-        light_state = self._read_by_identifier(id, IDENTIFIER_LIGHT_STATE)
-        belt = self._read_by_identifier(id, IDENTIFIER_BELT_STATE)
-        windows_closed = self._read_by_identifier(id, IDENTIFIER_WINDOWS_CLOSED)
-        data = [Door, serial_number, cigarette_lighter_voltage, light_state, belt, windows_closed]
+            Door = self._read_by_identifier(id, IDENTIFIER_DOOR)
+            serial_number = self._read_by_identifier(id, IDENTIFIER_DOOR_SERIALNUMBER)
+            cigarette_lighter_voltage = self._read_by_identifier(id, IDENTIFIER_LIGHTER_VOLTAGE)
+            light_state = self._read_by_identifier(id, IDENTIFIER_LIGHT_STATE)
+            belt = self._read_by_identifier(id, IDENTIFIER_BELT_STATE)
+            windows_closed = self._read_by_identifier(id, IDENTIFIER_WINDOWS_CLOSED)
+            data = [Door, serial_number, cigarette_lighter_voltage, light_state, belt, windows_closed]
 
-        module = DoorsToJSON()
-        response = module._to_json(data)
-        # Shutdown the CAN bus interface
-        self.bus.shutdown()
+            module = DoorsToJSON()
+            response = module._to_json(data)
+            # Shutdown the CAN bus interface
+            self.bus.shutdown()
 
-        log_info_message(logger, "Sending JSON")
-        return response
-    except CustomError as e:
-        self.bus.shutdown()
-        return e.message
+            log_info_message(logger, "Sending JSON")
+            return response
+        except CustomError as e:
+            self.bus.shutdown()
+            return e.message
 
     # def _to_json(self, module: ToJSON, data: list):
     #     """
