+++ src/rest_api/templates/index.html
@@ -44,7 +44,6 @@
             <button type="button" class="btn btn-info mb-3" onclick="readInfoEngine()">Read Info Engine</button>
             <button type="button" class="btn btn-info mb-3" onclick="readInfoDoors()">Read Info Doors</button>
             <button type="button" class="btn btn-info mb-3" onclick="writeInfoDoors()">Write Doors Info</button>
-            
         </div>
 
         <div class="mb-4">
